[GUI/CORE] Move protocol parser to the corelib
