+++ owmods_core/Cargo.toml
@@ -36,6 +36,7 @@ typeshare = "1.0.1"
 lazy_static = "1.4.0"
 tempfile = "3.7.1"
 unicode-normalization = "0.1.22"
+regex = "1.9.3"
 
 [dev-dependencies]
 tokio-test = "0.4.2"
